+++ test/images/agnhost/dns/common.go
@@ -17,10 +17,8 @@ limitations under the License.
 package dns
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
-	"os/exec"
 	"strings"
 
 	"github.com/spf13/cobra"
@@ -74,17 +72,4 @@ func readFile(fileName string) string {
 	}
 
 	return string(fileData)
-}
-
-func runCommand(name string, arg ...string) string {
-	var out bytes.Buffer
-	cmd := exec.Command(name, arg...)
-	cmd.Stdout = &out
-
-	err := cmd.Run()
-	if err != nil {
-		panic(err)
-	}
-
-	return strings.TrimSpace(out.String())
 }
