+++ test/images/agnhost/net/nat/closewait.go
@@ -38,6 +38,7 @@ import (
 
 // leakedConnection is a global variable that should leak the active
 // connection assigned here.
+//lint:ignore U1000 intentional unused variable
 var leakedConnection *net.TCPConn
 
 // CloseWaitServerOptions holds server JSON options.
