+++ test/images/agnhost/pause/pause.go
@@ -40,7 +40,6 @@ func pause(cmd *cobra.Command, args []string) {
 	done := make(chan int, 1)
 	signal.Notify(sigCh, syscall.SIGINT)
 	signal.Notify(sigCh, syscall.SIGTERM)
-	signal.Notify(sigCh, syscall.SIGKILL)
 	go func() {
 		sig := <-sigCh
 		switch sig {
@@ -50,9 +49,6 @@ func pause(cmd *cobra.Command, args []string) {
 		case syscall.SIGTERM:
 			done <- 2
 			os.Exit(2)
-		case syscall.SIGKILL:
-			done <- 0
-			os.Exit(0)
 		}
 	}()
 	result := <-done
